// packages/unlighthouse-old-mono/src/types.ts
export interface RouteDefinition {
  name: string
  path: string
  file?: string
  component?: string
  componentBaseName?: string
  chunkName?: string
  _name?: string
  layout?: string
}

/**
 * A normalised route, in the context of unlighthouse, is a mapping between a URL path, it's definition and a unique id.
 */
export interface NormalisedRoute {
  id: string
  path: string
  url: string
  $url: URL
  definition: RouteDefinition
  /**
   * A runtime path that the route was discovered from, useful if the route is a 404 and we want to know what directed
   * us to it.
   */
  discoveredFrom?: string
}

export interface ComputedLighthouseReportAudit {
  details?: {
    items?: any[]
  }
  displayValue: string | number
  score: number
}
/**
 * An augmented Lighthouse Report type, we add custom types to the base report for specific functionality on the
 * @unlighthouse/client.
 */
export type LighthouseReport = Partial<Result> & {
  /**
   * The total score for the result, this is the sum of each category's result
   */
  score: number
  categories: { score: number | null }[]
  computed: {
    /**
     * An aggregation of multiple image audit results.
     */
    imageIssues: ComputedLighthouseReportAudit
    ariaIssues: ComputedLighthouseReportAudit
  }
}

/**
 * Tasks that Unlighthouse will run, used to track their status.
 */
export type UnlighthouseTask = 'inspectHtmlTask' | 'runLighthouseTask'

/**
 * Each task ran by unlighthouse (extractHtmlPayload, runLighthouseTask) has a specific status which we can expose.
 */
export type UnlighthouseTaskStatus = 'waiting' | 'in-progress' | 'completed' | 'failed' | 'ignore' | 'failed-retry'

/**
 * A fairly rigid representation of the puppeteer cluster task results (extractHtmlPayload, runLighthouseTask), combined
 * with the normalised route.
 */
export interface UnlighthouseRouteReport {
  /**
   * The mapping of tasks with their status.
   */
  tasks: Record<UnlighthouseTask, UnlighthouseTaskStatus>
  /**
   * Track how long tasks are taking
   */
  tasksTime?: Record<Partial<UnlighthouseTask>, number>
  /**
   * Path to where the artifacts from a URL scan are saved.
   */
  artifactPath: string
  /**
   * URL of where the artifacts are stored, for static client access.
   */
  artifactUrl: string
  /**
   * The route (URL Path) that the report belongs to.
   */
  route: NormalisedRoute
  /**
   * A unique representation of the route, useful for the API layer.
   */
  reportId: string
  /**
   * The lighthouse result, only set once the task is completed.
   */
  report?: LighthouseReport
  /**
   * The SEO meta-data, only set once the html payload has been extracted and passed.
   */
  seo?: HTMLExtractPayload
}

export interface HTMLExtractPayload {
  alternativeLangDefault?: string
  alternativeLangDefaultHtml?: string
  title?: string
  description?: string
  internalLinks?: number
  externalLinks?: number
  htmlSize?: number
  favicon?: string
  og?: {
    description?: string
    title?: string
    image?: string
  }
}
// ...
export interface UnlighthouseWorkerStats {
  /**
   * Status of the worker, completed when all tasks have been completed.
   */
  status: 'completed' | 'working'
  /**
   * Time in ms that the worker has been running
   */
  timeRunning: number
  /**
   * How many tasks have been completed.
   */
  doneTargets: number
  /**
   * Total number of tasks including completed, pending and working.
   */
  allTargets: number
  /**
   * The % of work completed.
   */
  donePercStr: string
  /**
   * The % of errors.
   */
  errorPerc: string
  /**
   * The remaining time until all tasks are completed.
   */
  timeRemaining: number
  /**
   * How many tasks per second are being processed.
   */
  pagesPerSecond: string
  /**
   * The devices CPU usage % out of 100
   */
  cpuUsage: string
  /**
   * The devices memory usage % out of 100
   */
  memoryUsage: string
  /**
   * How many workers are now working, usually the cpu count of the device.
   */
  workers: number
}

export type PuppeteerTaskArgs = UnlighthouseRouteReport
export type PuppeteerTaskReturn = UnlighthouseRouteReport
// ...
export interface UnlighthouseWorker {
  /**
   * puppeteer-cluster instance
   */
  cluster: Cluster
  /**
   * A collection of stats gathered from the cluster for the current status of the worker.
   */
  monitor: () => UnlighthouseWorkerStats
  /**
   * Queue a single normalised route. Will not process routes that have already been queued.
   * @param route
   */
  queueRoute: (route: NormalisedRoute) => void
  /**
   * Queue multiple normalised routes. This will sort the list for a better loading experience.
   * @param routes
   */
  queueRoutes: (routes: NormalisedRoute[]) => void
  /**
   * Re-queues a report, avoiding the usual caching involved and makes sure we unlink any of the previous reports data or
   * tasks.
   *
   * @param report
   */
  requeueReport: (report: UnlighthouseRouteReport) => void
  /**
   * Has the worker started processing the queue.
   */
  hasStarted: () => boolean
  /**
   * The gathered map of reports. The key is the path of the route.
   */
  routeReports: Map<string, UnlighthouseRouteReport>
  /**
   * A simple array representation of the reports for easy iteration.
   */
  reports: () => UnlighthouseRouteReport[]
  /**
   * Find a report with the specified id.
   * @param id
   */
  findReport: (id: string) => UnlighthouseRouteReport | null

  /**
   * Iterates through route reports checking for a match on the route definition component, if there is a match
   * then the route is re-queued.
   *
   * @param file
   * @return True if an invalidation occurred on the routes.
   */
  invalidateFile: (file: string) => boolean

  /**
   * Checks if the amount of routes scanned surpasses the `scanner.maxRoutes` value.
   */
  exceededMaxRoutes: () => boolean

  /**
   * Clear the progress display box
   */
  clearProgressDisplay: () => void
}